fix(FGARetriever): Change key to relate FGA from docs
There are scenarios under which a single document source would split in
multiple chunks. When this happens, multiple documents will now have the
exact same FGA relation, thus, the FGA relation cannot be the key used
to retrieve the documents back from the FGA response. A new map is now
built, with the document as the key, as it is the ultimate identifier of
the relationships.